updating the update page to mdx files
