+++ src/updates/2023-06-29_Mobile-Communication-and-Online-Falsehoods-in-Asia.mdx
@@ -1,0 +1,7 @@
+---
+date: 2023-06-29
+title: "Mobile Communication and Online Falsehoods in Asia"
+excerpt: "The chapter written by Denny, Tarunima and Aditya speaks about the unique technical and social affordances of Mobile Instant Messaging Services (like WhatsApp) which, if accounted for, can result in a more productive agenda for protecting Indian citizens on these platforms"
+url: "https://link.springer.com/chapter/10.1007/978-94-024-2225-2_14"
+tags: Release
+---
