+++ src/updates/2023-06-30_DPG-Open-Source-Community-Manager-Programme.mdx
@@ -1,0 +1,7 @@
+---
+date: 2023-06-30
+title: "DPG Open Source Community Manager Programme"
+excerpt: "Uli took part in GitHub's efforts to popularise contributions to tech projects that work for the social good"
+url: "https://socialimpact.github.com/tech-for-social-good/dpg-open-source-community-manager-program"
+tags: Media Mentions
+---
