+++ src/updates/2023-07-31_Keynote-at-ADMS-Symposium-Alternative-Visions-for-Tech-in-the-Age-of-AI.mdx
@@ -1,0 +1,7 @@
+---
+date: 2023-07-01
+title: "Keynote at ADMS Symposium: Alternative Visions for Tech in the Age of AI"
+excerpt: "The talk focused on the importance of civic tech to healthier digital futures"
+url: "https://twitter.com/AdmsCentre/status/1679277464797995011"
+tags: Talk
+---
