--- src/updates/temp.mdx
@@ -1,7 +1,0 @@
----
-date: 2023-08-02
-title: "1"
-excerpt: "One"
-url: https://twitter.com/AdmsCentre/status/1679277464797995011
-tags: one, two
----
