--- src/updates/temp2.mdx
@@ -1,7 +1,0 @@
----
-date: 2024-07-02
-title: "2"
-excerpt: "Two"
-url: https://twitter.com/AdmsCentre/status/1679277464797995011
-tags: a,b,c,d
----
