--- src/updates/temp3.mdx
@@ -1,7 +1,0 @@
----
-date: 2023-10-04
-title: "3"
-excerpt: "Three"
-url: https://twitter.com/AdmsCentre/status/1679277464797995011
-tags: 1,2,3
----
